+++ src/Umbraco.Web.UI.Client/src/packages/core/section/conditions/section-alias/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_SECTION_ALIAS_CONDITION_ALIAS = 'Umb.Condition.SectionAlias';
