+++ src/Umbraco.Web.UI.Client/src/packages/core/section/conditions/section-alias/section-alias.condition.ts
@@ -1,6 +1,6 @@
-import { UmbConditionBase } from '../../extension-registry/conditions/condition-base.controller.js';
-import { UMB_SECTION_CONTEXT } from '../section.context.js';
-import type { SectionAliasConditionConfig } from './types.js';
+import { UmbConditionBase } from '../../../extension-registry/conditions/condition-base.controller.js';
+import { UMB_SECTION_CONTEXT } from '../../section.context.js';
+import type { SectionAliasConditionConfig } from '../types.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbConditionControllerArguments, UmbExtensionCondition } from '@umbraco-cms/backoffice/extension-api';
