+++ src/Umbraco.Web.UI.Client/src/packages/core/section/conditions/section-user-permission/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_SECTION_USER_PERMISSION_CONDITION_ALIAS = 'Umb.Condition.SectionUserPermission';
