+++ src/Umbraco.Web.UI.Client/src/packages/core/section/conditions/section-user-permission/types.ts
@@ -1,0 +1,20 @@
+import type { UMB_SECTION_USER_PERMISSION_CONDITION_ALIAS } from './constants.js';
+import type { UmbConditionConfigBase } from '@umbraco-cms/backoffice/extension-api';
+
+export type UmbSectionUserPermissionConditionConfig = UmbConditionConfigBase<
+	typeof UMB_SECTION_USER_PERMISSION_CONDITION_ALIAS
+> & {
+	/**
+	 *
+	 *
+	 * @example
+	 * "Umb.Section.Content"
+	 */
+	match: string;
+};
+
+declare global {
+	interface UmbExtensionConditionConfigMap {
+		UmbSectionUserPermissionConditionConfig: UmbSectionUserPermissionConditionConfig;
+	}
+}
